c++ project one README
Summarize the project and what problem it was solving.
This project was to create a program that read a text file to track items sold through out the day at a grocery store to see what was selling the most and how they should rearrange their store for  a better shopping experience. 

What did you do particularly well?
I created exactly what they asked for. A menu with 4 options:
1: Prompt a user to input the item, or word, they wish to look for.  Return a numeric value for the frequency of the specific word.
2: Print the list with numbers that represent the frequency of all items purchased.
3:  Print the same frequency information for all the items in the form of a histogram. Then print the name, followed by asterisks or another special character to represent the numeric amount.
4: Exit the program. 

Where could you enhance your code? How would these improvements make your code more efficient, secure, and so on?
User input validation: When getting user input, validate the input to ensure that it is within the acceptable range of values. This will prevent unexpected behavior or crashes due to invalid input. 

Which pieces of the code did you find most challenging to write, and how did you overcome this? What tools or resources are you adding to your support network?
I have never tried to pull from a text file before so this was completely new to me. I watched  some YouTube videos to help guide me on how to do it. When I don't know how or not understanding a concept I go straight to YouTube to see how it is done. Then I follow along or try break it down into what I am working on.  

What skills from this project will be particularly transferable to other projects or course work?
The main skill I think for me is pulling from a text file. That will help me when working with data as data can be stored in text files, so being able to write code to pull and write to a text file is an essential skill to have. 

How did you make this program maintainable, readable, and adaptable?
The program's maintainability and readability are enhanced by its organized structure, including the use of the Inventory class and function encapsulation. Descriptive variable and method names, as well as comments, make it easier for developers to understand the code. Standard libraries are used for common tasks, ensuring reliability and efficiency.

The adaptability of the program is facilitated by the modular design of the Inventory class. This allows new functionality to be added with minimal impact on existing code. Furthermore, methods within the class can be modified independently, ensuring better maintainability.